Split upload and download to separated files
