+++ pkg/keboola/storage_file.go
@@ -1,17 +1,13 @@
 package keboola
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
 	"sort"
 	"strconv"
 
 	"github.com/relvacode/iso8601"
-	"gocloud.dev/blob"
 
 	"github.com/keboola/go-client/pkg/client"
 	"github.com/keboola/go-client/pkg/keboola/storage_file_upload/abs"
@@ -256,139 +252,6 @@ func (a *API) DeleteFileRequest(id int) client.APIRequest[client.NoResult] {
 		}).
 		AndPathParam("fileId", strconv.Itoa(id))
 	return client.NewAPIRequest(client.NoResult{}, request)
-}
-
-type uploadConfig struct {
-	transport http.RoundTripper
-}
-
-type UploadOption func(c *uploadConfig)
-
-func WithUploadTransport(transport http.RoundTripper) UploadOption {
-	return func(c *uploadConfig) {
-		c.transport = transport
-	}
-}
-
-type downloadConfig struct {
-	transport http.RoundTripper
-}
-
-type DownloadOption func(c *downloadConfig)
-
-func WithDownloadTransport(transport http.RoundTripper) DownloadOption {
-	return func(c *downloadConfig) {
-		c.transport = transport
-	}
-}
-
-// NewUploadWriter instantiates a Writer to the Storage given by cloud provider specified in the File resource.
-func NewUploadWriter(ctx context.Context, file *FileUploadCredentials, opts ...UploadOption) (*blob.Writer, error) {
-	return NewUploadSliceWriter(ctx, file, "", opts...)
-}
-
-// NewUploadSliceWriter instantiates a Writer to the Storage given by cloud provider specified in the File resource and to the specified slice.
-func NewUploadSliceWriter(ctx context.Context, file *FileUploadCredentials, slice string, opts ...UploadOption) (*blob.Writer, error) {
-	c := uploadConfig{}
-	for _, opt := range opts {
-		opt(&c)
-	}
-	switch file.Provider {
-	case abs.Provider:
-		return abs.NewUploadWriter(ctx, file.ABSUploadParams, slice, c.transport)
-	case gcs.Provider:
-		return gcs.NewUploadWriter(ctx, file.GCSUploadParams, slice, c.transport)
-	case s3.Provider:
-		return s3.NewUploadWriter(ctx, file.S3UploadParams, file.Region, slice, c.transport)
-	default:
-		return nil, fmt.Errorf(`unsupported provider "%s"`, file.Provider)
-	}
-}
-
-// Upload instantiates a Writer to the Storage given by cloud provider specified in the File resource and writes there
-// content of the reader.
-func Upload(ctx context.Context, file *FileUploadCredentials, fr io.Reader) (written int64, err error) {
-	return UploadSlice(ctx, file, "", fr)
-}
-
-// UploadSlice instantiates a Writer to the Storage given by cloud provider specified in the File resource and writes
-// content of the reader to the specified slice.
-func UploadSlice(ctx context.Context, file *FileUploadCredentials, slice string, fr io.Reader) (written int64, err error) {
-	bw, err := NewUploadSliceWriter(ctx, file, slice)
-	if err != nil {
-		return 0, fmt.Errorf("cannot open bucket writer: %w", err)
-	}
-
-	defer func() {
-		if closeErr := bw.Close(); closeErr != nil && err == nil {
-			err = fmt.Errorf("cannot close bucket writer: %w", closeErr)
-		}
-	}()
-
-	return io.Copy(bw, fr)
-}
-
-// UploadSlicedFileManifest instantiates a Writer to the Storage given by cloud provider specified in the File resource and writes
-// content of the reader to the specified slice manifest.
-func UploadSlicedFileManifest(ctx context.Context, file *FileUploadCredentials, slices []string) (written int64, err error) {
-	manifest, err := NewSlicedFileManifest(file, slices)
-	if err != nil {
-		return 0, err
-	}
-	marshaledManifest, err := json.Marshal(manifest)
-	if err != nil {
-		return 0, err
-	}
-
-	return UploadSlice(ctx, file, ManifestFileName, bytes.NewReader(marshaledManifest))
-}
-
-func Download(ctx context.Context, file *FileDownloadCredentials) ([]byte, error) {
-	return DownloadSlice(ctx, file, "")
-}
-
-func DownloadManifest(ctx context.Context, file *FileDownloadCredentials) ([]byte, error) {
-	return DownloadSlice(ctx, file, ManifestFileName)
-}
-
-func DownloadSlice(ctx context.Context, file *FileDownloadCredentials, slice string) (out []byte, err error) {
-	reader, err := DownloadSliceReader(ctx, file, slice)
-	if err != nil {
-		return nil, err
-	}
-	out, err = io.ReadAll(reader)
-	if closeErr := reader.Close(); err == nil && closeErr != nil {
-		err = closeErr
-	}
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func DownloadReader(ctx context.Context, file *FileDownloadCredentials) (io.ReadCloser, error) {
-	return DownloadSliceReader(ctx, file, "")
-}
-
-func DownloadManifestReader(ctx context.Context, file *FileDownloadCredentials) (io.ReadCloser, error) {
-	return DownloadSliceReader(ctx, file, ManifestFileName)
-}
-
-func DownloadSliceReader(ctx context.Context, file *FileDownloadCredentials, slice string, opts ...DownloadOption) (io.ReadCloser, error) {
-	c := downloadConfig{}
-	for _, opt := range opts {
-		opt(&c)
-	}
-	switch file.Provider {
-	case abs.Provider:
-		return abs.NewDownloadReader(ctx, file.ABSDownloadParams, slice, c.transport)
-	case gcs.Provider:
-		return gcs.NewDownloadReader(ctx, file.GCSDownloadParams, slice, c.transport)
-	case s3.Provider:
-		return s3.NewDownloadReader(ctx, file.S3DownloadParams, file.Region, slice, c.transport)
-	default:
-		return nil, fmt.Errorf(`unsupported provider "%s"`, file.Provider)
-	}
 }
 
 func NewSliceURL(file *FileUploadCredentials, slice string) (string, error) {
